+++ examples/Rules-method-size-CohortSizeParts.R
@@ -32,16 +32,16 @@ nextMaxDose <- maxDose(myIncrements,
 # based on the class 'NextBestNCRM'
 myNextBest <- NextBestNCRM(target=c(0.2, 0.35),
                            overdose=c(0.35, 1),
-                           maxOverdoseProb=0.25)
+                           max_overdose_prob=0.25)
 
 # Calculate the next best dose
 doseRecommendation <- nextBest(myNextBest,
                                doselimit=nextMaxDose,
-                               samples=samples, 
-                               model=model, 
+                               samples=samples,
+                               model=model,
                                data=data)
 
-# Rule for the cohort size: 
+# Rule for the cohort size:
 mySize <- CohortSizeParts(sizes=c(1,3))
 
 # Determine the cohort size for the next cohort