+++ examples/Rules-method-stopTrial-StoppingMinCohorts.R
@@ -32,19 +32,19 @@ nextMaxDose <- maxDose(myIncrements,
 # based on the class 'NextBestNCRM'
 myNextBest <- NextBestNCRM(target=c(0.2, 0.35),
                            overdose=c(0.35, 1),
-                           maxOverdoseProb=0.25)
+                           max_overdose_prob=0.25)
 
 # Calculate the next best dose
 doseRecommendation <- nextBest(myNextBest,
                                doselimit=nextMaxDose,
                                samples=samples, model=model, data=data)
 
-# Define the stopping rule such that the study would be stopped if at least 6 
-# cohorts were already dosed 
+# Define the stopping rule such that the study would be stopped if at least 6
+# cohorts were already dosed
 myStopping <- StoppingMinCohorts(nCohorts = 6)
 
 # Evaluate if to stop the trial
-stopTrial(stopping=myStopping, 
+stopTrial(stopping=myStopping,
           dose=doseRecommendation$value,
           data=data)